backend/requirements.txt
fastapi
uvicorn
pydantic
celery
redis
websockets
psycopg2-binary
google-genai
rich
